pyproject.toml
[tool.poetry]
name = "viqzo"
version = "0.1.0"
description = ""
authors = ["HETPAHEP"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.2"
djangorestframework = "3.14.0"
python-dotenv = "1.0.0"
pytelegrambotapi = "4.12.0"

[tool.poetry.group.dev.dependencies]
mypy = "1.4.1"
isort = "5.12.0"
flake8 = "6.0.0"
pytest = "7.4.0"
pytest-django = "4.5.2"
coverage = "7.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
